ServiceConnector
This app let’s Service providers such as Plumbers, Carpenters and all
connect to people needed the service right from their mobile phones.
In this Application the user post for a job request stating the
resource he needs. This sends a notification to person involved with
that particular skill set in vicinity of the area.